Move contrast change to separate file and set slit hgaps correctly in base
